# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "URBIS API",
    "version": "0.5.0-alpha.1",
    "description": "API da plataforma URBIS. Propriedade: Sigma SIHF Soluções Analíticas S/A — CNPJ 01.851.824/0001-38. A especificação diferencia funcionalidades implementadas de dependências externas que ainda exigem validação em staging."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Backend NestJS local"
    },
    {
      "url": "http://localhost:3100",
      "description": "Prova local em memória"
    }
  ],
...
  "paths": {
...
    "/v1/territory/cep/{cep}": {
      "get": {
        "summary": "Consultar CEP",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Territory"
        ],
        "security": [],
        "parameters": [
          {
            "name": "cep",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/territory/nearest": {
      "get": {
        "summary": "Logradouros próximos a um ponto",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Territory"
        ],
        "security": [],
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "lon",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/territory/streets": {
      "get": {
        "summary": "Segmentos GIS por bounding box",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Territory"
        ],
        "security": [],
        "parameters": [
          {
            "name": "minLon",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "minLat",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxLon",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxLat",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
...
}